BaseUtils has more clear distinction between byte and char routines.  All char routines are @Depreciated now.  Please use bytes.  Better organization of reverse(), now in Utils not BaseUtils.
git-svn-id: file:///humgen/gsa-scr1/gsa-engineering/svn_contents/trunk@3400 348d0f76-0448-11de-a6fe-93d51630548a

+++ java/src/org/broadinstitute/sting/oneoffprojects/walkers/ReadErrorRateWalker.java
@@ -100,18 +100,18 @@ public class ReadErrorRateWalker extends ReadWalker<boolean[], ReadErrorRateColl
                 if (useNextBestBase || useNextRandomBase || useNonNextBestBase) {
                     byte nextBestBase;
                     if (useNextBestBase) {
-                        nextBestBase = (byte) BaseUtils.baseIndexToSimpleBase(QualityUtils.compressedQualityToBaseIndex(sq[cycle]));
+                        nextBestBase = BaseUtils.baseIndexToSimpleBase(QualityUtils.compressedQualityToBaseIndex(sq[cycle]));
                     } else if (useNonNextBestBase) {
                         nextBestBase = bases[cycle];
                         Random generator = new Random();
-                        while (nextBestBase == bases[cycle] || nextBestBase == (byte) BaseUtils.baseIndexToSimpleBase(QualityUtils.compressedQualityToBaseIndex(sq[cycle]))) {
-                            nextBestBase = (byte) BaseUtils.baseIndexToSimpleBase(generator.nextInt(4));
+                        while (nextBestBase == bases[cycle] || nextBestBase == BaseUtils.baseIndexToSimpleBase(QualityUtils.compressedQualityToBaseIndex(sq[cycle]))) {
+                            nextBestBase = BaseUtils.baseIndexToSimpleBase(generator.nextInt(4));
                         }
                     } else {
                         nextBestBase = bases[cycle];
                         Random generator = new Random();
                         while (nextBestBase == bases[cycle]) {
-                            nextBestBase = (byte) BaseUtils.baseIndexToSimpleBase(generator.nextInt(4));
+                            nextBestBase = BaseUtils.baseIndexToSimpleBase(generator.nextInt(4));
                         }
                     }
